MVP Instancing V1
This commit is the first to provide a needed features to serve
instanced problems.

Starting with this commit, the admin can use Jinja2 templating
in the problem name, description, and keys. The rendering
parameters are specified as a json dictionary under the
"Instances" section, with each parameters set becoming a
instance which will be served at "random" to users