POC of Embedded Shell
This is a quick Proof of Concept for an embedded shell in the
browser using shellinabox. To use startup CTFd and login as a
regular user and navigate to the shell tab. If it is not working
make sure you have shellinabox installed and that it is running
on port 4200